finance-chatbot/requirements.txt
flask
flask-cors
requests
waitress
haystack-ai
